requirements.txt
# Dependencies for GuppShupp Memory & Personality Engine
groq>=0.4.0
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.26.0
